catch timeouts
should resolve #1
when the applicationmetadata is null then there is no application
connected, usually because the network connection has gone away or the
cast session has sat idle for too long.
If this timeout occurs update the javascript side of the plugin to
disconnected state and alert the user of the device that has timed out
so they can take appropriate action.