update launch and join application
by default launch will start with a clean app state, so don’t have to
worry about what may have been running on the receiver previously.
update the join method to join if the current application is the same
as the requested application else relaunch the requested application on
device.