Update framework & install scripts
Google cast framework is no longer statically embedded as it is now
HUGE due to bit code inclusion apparently. End app install size is not
reflective of this but makes it impossible to commit the framework.

Added install and uninstall scripts
scripts currently hard set to download the latest castSDK version as of
today which is v3.1.1 and install automatically.